requirements.txt
Django==4.1.5
psycopg2==2.9.5
python-decouple==3.7
pytz==2022.7.1
djangorestframework==3.14.0